Fix: Refactor MultiSigPageSetup component
Fix: Refactor MultiSigPageSetup component
